// delis-angular/src/app/model/content/document/invoice-response-generation.model.ts
export class InvoiceResponseGenerationModel {

  status: string;
  action: string;
  actionEnabled: boolean;
  action2: string;
  action2Enabled: boolean;
  reason: string;
  reasonEnabled: boolean;
  detailType: string;
  detailValue: string;
  statusReasonText: string;
  effectiveDate: string;
  effectiveDateEnabled: boolean;
}